Consolidate exports into single file
