--- src/euidExports.ts
@@ -1,2 +1,0 @@
-export { EventType, Uid2CallbackHandler } from './uid2CallbackManager';
-export { CallbackContainer, ProductDetails, UID2SdkBase, UID2Setup } from './sdkBase';
