+++ src/euidSdk.ts
@@ -1,7 +1,7 @@
 import { EventType, Uid2CallbackHandler } from './uid2CallbackManager';
 import { CallbackContainer, ProductDetails, UID2SdkBase, UID2Setup } from './sdkBase';
 
-export * from './euidExports';
+export * from './exports';
 
 export class EUID extends UID2SdkBase {
   private static cookieName = '__euid';
